[yeojiin(임정연)] - 4단계 구간 제거 기능 (#980)
* doc(README.md): 문서 수정

* doc(README.md): 2단계 미션 문서 정리

* fix(StationService): 단수형 명칭 수정

* fix(StationService): 단수형 명칭 수정

* refactor(StationAcceptanceTest): 한글 메소드로 변경

* feat(LineController): 노선 생성 구현

* test(LineAcceptanceTest): 노선 생성 인수 테스트 구현 -> 성공

* test(LineAcceptanceTest): 노선 목록 조회 구현 -> 성공

* test(LineAcceptanceTest): 노선 조회 구현 -> 성공

* test(LineAcceptanceTest): 노선 삭제 -> 성공

* test(LineAcceptanceTest): 노선 정보 수정 -> 색상 변경 오류

* fix(LineResponse): color에 name리턴 수정

* doc(README.md): 기능 구현 체크

* doc(README.md): 3단계 요구사항 목록 정리

* fix(LineAcceptanceStep): 메서드화 작업이 많아져 클래스 분리 -> 코멘트 반영

* fix(LineService): LineReponse.of 메서드 생성

* feat(Section): Section, Sections 생성

* fix(LineAcceptanceTest): Line 수정된 내용에 맞게 테스트케이스 수정

* feat(Sections): 구간 등록 시 validation 구현중

* feat(Sections): 구간등록 시 기존 역, 거리 정보 수정 구현

* feat(Sections): 구간등록 시 기존 역, 거리 정보 수정 구현

* fix(Line): SectionRepository는 삭제 -> section은 line에 저장하면 되기 때문에
Section을 Line과 연결하기 위해 생성자 파라미터 변경

* test(LineSectionTest): 구간등록, 중복구간등록 에러 성공

* test(LineSectionTest): 구간 등록 중 상행,하행, 하나라도 등록되어있어야 가능 테스트 -> 성공

* test(LineSectionTest): 새로운 구간 길이 제한 테스트 -> 성공

* refactor(LineSectionTest): 공통 호출 메소드 리팩토링

* test(LineSectionTest): 상행 종점 등록, Line Response에 sections 추가

* fix(Sections): 상행,종행 종점일 경우 길이 체크하지 않는 로직 구현중-> compareTo Override

* feat(Sections): sort 작업 진행중 중단

* refactor(Sections): 주석 정리 및 메소드 정리

* doc(README.md)：기능목록 작성

* doc(README.md): doc 수정

* refactor(module): 정리

* refactor(all): updateUpStation으로 stream 사용해서 리펙토링

* refactor(LineResponse): 구간 등록 시 lineResponse.section sort해서 순서대로 보기

* feat(LineController): 역 삭제 구현중

* refactor(Section): 오타수정

* refactor(Section): 롤백

* fix(module): step3 merge

* refactor(test): CommonMethodFixture로 공통 Method 추출

* fix(Section): @ManyToOne Lazy 삭제

* test(LineAcceptanceTest): 종점 구간 삭제 테스트

* test(LineAcceptanceTest): 중간 구간 삭제, 노선이 하나인 구간 삭제 테스트

* refactor(Sections): 구간 삭제 리팩토링

* refactor(LineResponse): LineReponse 불필요한 메소드 정리

* refactor(LineResponse): LineReponse 불필요한 메소드 정리

* refactor(LineAcceptamceStep): 메소드 정리

* fix(ControllerExceptionHandler): exception handler -> 400 return

* fix(LineService): boxing 수정

* fix(LineSectionTest): Long 수정

* fix(LineResponse): Section 삭제

* fix(LineResponse): Section 삭제

* refactor(LineSectionStep): 에러메세지 확인 추가

* feat(LineSectionTest): 해당 구간이 아닌 역 삭제 추가

* feat(LineSectionTest): 해당 구간이 아닌 역 삭제 추가

+++ src/main/java/nextstep/subway/dto/LineRequest.java
@@ -21,9 +21,9 @@ public class LineRequest {
     private String color;
     private int distance;
     @Builder.Default
-    private int upLastStationId = 0;
+    private Long upLastStationId = 0L;
     @Builder.Default
-    private int downLastStationId = 0;
+    private Long downLastStationId = 0L;
 
     public Line toLine(Station upStation, Station downStation) {
         return new Line(name, color, upStation, downStation, distance);
